requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
httpx
structlog
slowapi
passlib[bcrypt]
